Update README.md
This game was written largely as an exercise in how to write a Mac GUI application using Java Swing.  It includes a 'model' that has no dependency on the GUI elements, and, in fact, runs via unit tests, outputting the game state as text.  It uses Java Swing to do the GUI.  It uses Maven to build the .jar and the assembly plugin to build a 'jar-with-dependencies'.  Finally, it uses an ant jarbundler task to bundle the resulting files into a standalone executable Mac OS X application that includes the java runtime environment so there are no external dependencies.